w3c-reports/12092014/mac/runned-tests/html/browsers/history/the-history-interface/002-expected.txt

WARNING: This test should always be loaded in a new tab/window, to avoid browsers attempting to recover the state of frames, and history length. Do not reload the test.

Running test...









FAIL history.length should update when loading pages in an iframe assert_equals: make sure that you loaded the test in a new tab/window expected 101 but got 100(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/browsers/history/the-history-interface/002.html:40:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
tests1@http://localhost:8800/html/browsers/history/the-history-interface/002.html:40:7)
FAIL history.length should update when setting location.hash assert_equals: make sure that you loaded the test in a new tab/window expected 101 but got 100(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/browsers/history/the-history-interface/002.html:44:17
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
tests1@http://localhost:8800/html/browsers/history/the-history-interface/002.html:43:7)
PASS history.replaceState must exist 
PASS history.replaceState must exist within iframes 
PASS initial history.state should be null 
PASS history.length should not update when replacing a state with no URL 
PASS history.state should update after a state is pushed 
PASS hash should not change when replaceState is called without a URL 
PASS history.length should not update when replacing a state with a URL 
PASS hash should change when replaceState is called with a URL 
PASS replaceState must replace the existing state and not add an extra one 
PASS replaceState must replace the existing state without altering the forward history 
PASS replaceState must not be allowed to create invalid URLs 
PASS replaceState must not be allowed to create cross-origin URLs 
PASS replaceState must not be allowed to create cross-origin URLs (about:blank) 
PASS replaceState must not be allowed to create cross-origin URLs (data:URI) 
PASS security errors are expected to be thrown in the context of the document that owns the history object 
PASS replaceState must be able to set location.pathname 
PASS replaceState must be able to set absolute URLs to the same host 
PASS replaceState must not remove any tasks queued by the history traversal task source 
PASS .go must queue a task with the history traversal task source (run asynchronously) 
PASS replaceState must not fire hashchange events 
PASS replaceState must not be able to use a function as data 
PASS replaceState must not be able to use a DOM node as data 
PASS replaceState must not be able to use an error object as data 
PASS security errors are expected to be thrown in the context of the document that owns the history object 
PASS replaceState must be able to make structured clones of complex objects 
PASS history.state should also reference a clone of the original object 
PASS history.state should be a clone of the original object, not a reference to it 
PASS popstate event should fire when navigation occurs 
PASS popstate event should pass the state data 
PASS state data should cope with circular object references 
PASS state data should be a clone of the original object, not a reference to it 
PASS history.state should also reference a clone of the original object 
PASS history.state should be a clone of the original object, not a reference to it 
FAIL history.state should be a separate clone of the object, not a reference to the object passed to the event handler assert_false: expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/html/browsers/history/the-history-interface/002.html:253:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
tests9@http://localhost:8800/html/browsers/history/the-history-interface/002.html:252:7
http://localhost:8800/html/browsers/history/the-history-interface/002.html:212:98)
PASS replaceState should not actually load the new URL 
PASS reloading a replaced state should actually load the new URL 
